+++ EpicJewels/GemEffects/CombatSpirit.cs
@@ -24,7 +24,7 @@ namespace EpicJewels.GemEffects
         {
             private static void Postfix(Player __instance, bool sensed, bool alerted)
             {
-                if (Player.m_localPlayer.GetEffectPower<Config>("Combat Spirit").Power > 0 && __instance == Player.m_localPlayer && sensed && alerted)
+                if (Player.m_localPlayer != null && Player.m_localPlayer.GetEffectPower<Config>("Combat Spirit").Power > 0 && __instance == Player.m_localPlayer && sensed && alerted)
                 {
                     if (wolf == null)
                     {