+++ EpicJewels/GemEffects/WaterSwiftness.cs
@@ -18,7 +18,7 @@ namespace EpicJewels.GemEffects
         {
             private static void Postfix(Player __instance, ref float __result)
             {
-                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && (__instance.GetEffectPower<Config>("Water Swiftness").Power > 0 || __instance.GetEffectPower<Config>("Slippery When Wet").Power > 0))
+                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) || __instance.GetEffectPower<Waterproof.Config>("Waterproof").Power > 0 && (__instance.GetEffectPower<Config>("Water Swiftness").Power > 0 || __instance.GetEffectPower<Config>("Slippery When Wet").Power > 0))
                 {
                     float speed_modifier = ((__instance.GetEffectPower<Config>("Water Swiftness").Power + __instance.GetEffectPower<Config>("Slippery When Wet").Power + 100f) / 100f);
                     // EpicJewels.EJLog.LogDebug($"Wet jog speed modifier: {speed_modifier}");
@@ -33,7 +33,7 @@ namespace EpicJewels.GemEffects
         {
             private static void Postfix(Player __instance, ref float __result)
             {
-                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && (__instance.GetEffectPower<Config>("Water Swiftness").Power > 0 || __instance.GetEffectPower<Config>("Slippery When Wet").Power > 0))
+                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) || __instance.GetEffectPower<Waterproof.Config>("Waterproof").Power > 0 && (__instance.GetEffectPower<Config>("Water Swiftness").Power > 0 || __instance.GetEffectPower<Config>("Slippery When Wet").Power > 0))
                 {
                     float speed_modifier = ((__instance.GetEffectPower<Config>("Water Swiftness").Power + __instance.GetEffectPower<Config>("Slippery When Wet").Power + 100f) / 100f);
                     // EpicJewels.EJLog.LogDebug($"Wet run speed modifier: {speed_modifier}");